refactor:descreibe log extend info
