+++ polaris-console.yaml
@@ -22,10 +22,9 @@ polarisServer:
   polarisToken: "polaris@12345678"
 monitorServer:
   address: "127.0.0.1:9090"
-eventServer:
-  requestUrl: 
-  timeout: 5s
-operationServer:
-  requestUrl:
-  timeout: 5s
-futures: ${POLARIS_FUTURES}
+# eventServer:
+#   requestUrl: 
+#   timeout: 5s
+# operationServer:
+#   requestUrl:
+#   timeout: 5s
